Revised sign-in flow, reorg utils
