+++ app/routes/account.js
@@ -4,6 +4,17 @@ const utils = require('../utils')
  * Account routes
  */
 module.exports = router => {
+  router.post('/account', (req, res) => {
+    const hasAccount = req.session.data['has-account']
+
+    if (hasAccount === 'yes') {
+      utils.sendEmail(req, 'c3457068-675e-4ff9-963e-2e7444607bad')
+      res.redirect('/account/check-email/sign-in')
+    } else {
+      res.redirect('/account/eligibility')
+    }
+  })
+
   router.get('/account/eligibility/answer', (req, res) => {
     const eligibileNationality = req.session.data['eligibile-nationality']
     const eligibileEquivalencies = req.session.data['eligibile-equivalencies']
