+++ app/routes/email.js
@@ -1,39 +1,4 @@
-const { DateTime } = require('luxon')
-const NotifyClient = require('notifications-node-client').NotifyClient
-const notify = new NotifyClient(process.env.NOTIFYAPIKEY)
-
-const providers = require('./../data/providers')
-
-// Emails will only send if the email address has been whitelisted
-const sendEmail = (req, template, personalisation) => {
-  const email = req.session.data.account && req.session.data.account.email
-  personalisation = personalisation || {}
-  personalisation.url = req.get('origin') || `${req.protocol}://${req.get('host')}`
-
-  if (email) {
-    notify.sendEmail(
-      template,
-      email,
-      { personalisation }
-    )
-  }
-}
-
-const getProvider = (providerCode) => {
-  return providers[providerCode]
-}
-
-const getCourse = (providerCode, courseCode) => {
-  return providers[providerCode].courses[courseCode]
-}
-
-const nowPlusDays = (days, format = 'yyyy-LL-dd') => {
-  const date = DateTime.local().plus({ days: days })
-
-  return DateTime.fromISO(date, {
-    locale: 'en-GB'
-  }).toFormat(format)
-}
+const utils = require('../utils')
 
 // UR: Hard-coded template values for use during research sesions
 const urChoices = [{
@@ -63,7 +28,7 @@ module.exports = router => {
   // Account: Welcome
   router.get('/account/account-created', (req, res) => {
     if (!req.session.data.welcomeEmailSent) {
-      sendEmail(req, '8aa601b6-9985-471e-bd28-5fe0697820e5')
+      utils.sendEmail(req, '8aa601b6-9985-471e-bd28-5fe0697820e5')
     }
 
     req.session.data.welcomeEmailSent = true
@@ -72,13 +37,13 @@ module.exports = router => {
 
   // Account: Sign up
   router.post('/send-email/create-account', (req, res) => {
-    sendEmail(req, 'd19c241f-73f9-45da-a9cc-6613f9b3d2da')
+    utils.sendEmail(req, 'd19c241f-73f9-45da-a9cc-6613f9b3d2da')
     res.redirect('/account/check-email/create-account')
   })
 
   // Account: Sign in
   router.post('/send-email/sign-in', (req, res) => {
-    sendEmail(req, 'c3457068-675e-4ff9-963e-2e7444607bad')
+    utils.sendEmail(req, 'c3457068-675e-4ff9-963e-2e7444607bad')
     res.redirect('/account/check-email/sign-in')
   })
 
@@ -91,8 +56,8 @@ module.exports = router => {
     const choices = []
     for (const choice in application.choices) {
       const { courseCode, providerCode } = application.choices[choice]
-      const provider = getProvider(providerCode)
-      const course = getCourse(providerCode, courseCode)
+      const provider = utils.getProvider(providerCode)
+      const course = utils.getCourse(providerCode, courseCode)
       const result = `* ${provider.name} - ${course.name_and_code}\nStarting 5 September 2020`
       choices.push(result)
     }
@@ -106,11 +71,11 @@ module.exports = router => {
       notifyTemplate = '99a20df5-564d-4612-810e-3788edf7285e'
     }
 
-    sendEmail(req, notifyTemplate, {
+    utils.sendEmail(req, notifyTemplate, {
       reference: applicationId,
       candidateName,
       choiceList: choices.join('\n'),
-      amendDate: nowPlusDays(7, 'd MMMM yyyy')
+      amendDate: utils.nowPlusDays(7, 'd MMMM yyyy')
     })
     res.redirect(`/application/${applicationId}/confirmation`)
   })
@@ -146,9 +111,9 @@ module.exports = router => {
       }
     }
 
-    sendEmail(req, notifyTemplate, {
+    utils.sendEmail(req, notifyTemplate, {
       candidateName: name || 'candidate',
-      dbdDate: nowPlusDays(12, 'd MMMM yyyy'),
+      dbdDate: utils.nowPlusDays(12, 'd MMMM yyyy'),
       providerName,
       courseName,
       courseDate,
@@ -198,7 +163,7 @@ module.exports = router => {
     }
 
     // UR: Decision is made on third course choice
-    sendEmail(req, notifyTemplate, {
+    utils.sendEmail(req, notifyTemplate, {
       reference: applicationId,
       candidateName: application['given-name'] || 'applicant',
       providerName: urChoices[2].providerName,
