+++ app/views/account/check-email.njk
@@ -5,7 +5,7 @@
 
 {% block primary %}
   {% if action == "sign-in"%}
-    <p><a class="app-hidden-link" href="/applications">We’ve sent a sign in link to {{ data.account.email or 'your-email@domain.com' }}</a>.</p>
+    <p><a class="app-hidden-link" href="/applications?token=true">We’ve sent a sign in link to {{ data.account.email or 'your-email@domain.com' }}</a>.</p>
 
     <p>If our email doesn’t arrive within 5 minutes, check your spam and trash folder, or <a href="/account/sign-in">try again</a>.</p>
   {% else %}
