+++ app/views/index.njk
@@ -36,12 +36,10 @@
 
   {{ govukButton({
     text: "Start now",
-    href: "/account/eligibility",
+    href: "/account",
     isStartButton: true,
     classes: "govuk-!-margin-top-4 govuk-!-margin-bottom-3"
   }) }}
-  <p class="govuk-body">or</p>
-  <p class="govuk-body"><a href="/account/sign-in">Sign in to an existing application</a></p>
 {% endblock %}
 
 {% block secondary %}
